+++ RELEASE-NOTES.md
@@ -27,6 +27,7 @@
 * #493 IntReverseSequenceSeries - decreasing sequential IntSeries
 * #494 Upgrade to Parquet 1.15.2
 * #495 Force column name deduplication in SqlLoader
+* #496 Force column name deduplication in CsvLoader
 
 ## Release 2.0.0-M2